move Formatter into fmt module
